refactor(app): co-locate the two ARIA-to-locator generators
generateFromAriaSnapshot (the page-wide relevance-scored fallback) moves
from server/utils/locator-healing.ts to shared/locator-fingerprint.ts, next
to its sibling freshLocatorsFromCandidate (the fixed-band scorer for a
single matched candidate), under a header that contrasts the two scoring
regimes so the divergence stops reading as an accident. Its private helpers
(extractFailingText, tokenize) move with it and it reuses the file's
existing escapeQuote. No scoring or behavior change — the server imports it
from #shared. Now exported and unit-tested (token-overlap and position-bonus
ordering, getByText/getByLabel emission).
